mm-video-v4l2: Fix compilation errors for 64 bit build
Fix compilation issues in decoder omx components.

+++ mm-video-v4l2/vidc/vdec/src/omx_vdec_msm8974.cpp
@@ -11205,8 +11205,8 @@ OMX_ERRORTYPE omx_vdec::describeColorFormat(DescribeColorFormatParams *params) {
     DEBUG_PRINT_LOW("  FrameWidth x FrameHeight : %d x %d", params->nFrameWidth, params->nFrameHeight);
     DEBUG_PRINT_LOW("  YWidth x YHeight : %d x %d", img->mWidth, img->mHeight);
     for (size_t i = 0; i < img->mNumPlanes; ++i) {
-        DEBUG_PRINT_LOW("    Plane[%d] : offset=%d / xStep=%d / yStep = %d",
-                i, img->mPlane[i].mOffset, img->mPlane[i].mColInc, img->mPlane[i].mRowInc);
+        DEBUG_PRINT_LOW("    Plane[%d] : offset=%u / xStep=%u / yStep = %u",
+                (int)i, img->mPlane[i].mOffset, img->mPlane[i].mColInc, img->mPlane[i].mRowInc);
     }
     return OMX_ErrorNone;
 }
